Rename Editor directory to editor in lowercase
