+++ components/Sidebar/index.tsx
@@ -19,7 +19,7 @@ import {
 } from '@heroicons/react/outline'
 
 import { useFetchAll } from 'lib/hooks/useFetchAll'
-import { Logo } from 'components/Logo'
+import { Logo } from 'components/common/Logo'
 
 const Sidebar = ({ show }: { show: boolean }) => {
   const { user, isLoading, error } = useUser()
