+++ pages/[id]/index.tsx
@@ -12,7 +12,7 @@ import { sidebarAtom } from 'pages/create'
 import { Tab } from '@headlessui/react'
 import { MenuIcon, PencilIcon } from '@heroicons/react/outline'
 import { useResponses } from 'lib/hooks/useResponses'
-import { LabelSwitch } from 'components/Editor/LabelSwitch'
+import { LabelSwitch } from 'components/editor/LabelSwitch'
 import { mutate } from 'swr'
 import toast from 'react-hot-toast'
 import { OverlayPage } from 'components/OverlayPage'
